+++ CHANGELOG.md
@@ -3,10 +3,11 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-[Unreleased]
+[4.60.5] - 2018-10-25
 ----------------------
 ##### Fixed
 - Autocomplete widget not updating when form language is changed.
+- Failing to translate select options from external data (Survey123 only).
 
 [4.60.4] - 2018-10-19
 ----------------------